% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{add_sample_xpos}
\alias{add_sample_xpos}
\title{Add a per-sample xpos column for sample data based on grouping}
\usage{
add_sample_xpos(data, group_cols, group_order = NULL)
}
\arguments{
\item{data}{a data.frame with both annotations and gene expression vales}

\item{group_cols}{a group_cols list generated by group_columns()}

\item{group_order}{an optional vector of group ids for ordering (default is NULL).}
}
\value{
a data.frame of data with a column xpos appended.
}
\description{
Add a per-sample xpos column for sample data based on grouping
}
