% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{auto_format}
\alias{auto_format}
\title{Title}
\usage{
auto_format(df, sample_id = NULL, anno_cols = NULL, data_cols = NULL)
}
\arguments{
\item{df}{}

\item{sample_id}{}

\item{anno_cols}{}

\item{data_cols}{}
}
\description{
Title
}
