% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{build_header_labels}
\alias{build_header_labels}
\title{Build colorful, rectangular labels for plot headers in plot space}
\usage{
build_header_labels(
  data,
  grouping,
  group_order = NULL,
  ymin,
  label_height = 25,
  label_type = "simple"
)
}
\arguments{
\item{data}{A data.frame containing joined annotations and expression data}

\item{grouping}{The base to use for grouping samples.}

\item{group_order}{Alternate order to use for grouping if necessary. Default is NULL, which will use grouping_id.}

\item{ymin}{The minimum y value for the bottom of the headers. In gene plot contexts, this will usually be the number of genes + 1}

\item{label_height}{Percentage of the plot area that the headers should take up. Default = 25.}

\item{label_type}{Either "simple", "angle", or "square". Simple is for use with grouped plots. 
Angle will scale the header bars to be equal width, with a layer of 
angled polygons connecting them to the data. Square will construct headers with rectangular bars that match the width 
of the samples in each group.}
}
\description{
Build colorful, rectangular labels for plot headers in plot space
}
