% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{build_max_dfs}
\alias{build_max_dfs}
\title{Build data.frames for maximum value positions in plot space}
\usage{
build_max_dfs(n_stats, max_vals, max_width)
}
\arguments{
\item{n_stats}{count statistics made by get_n_stats()}

\item{max_vals}{maximum values per gene generated by max_gene_vals()}

\item{max_width}{the percentage of the plot that the max values should occupy.}
}
\value{
a list with 3 data.frames: labels with label positions and values; header, with header position and label; width with the width in plot space.
}
\description{
Build data.frames for maximum value positions in plot space
}
