% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{build_river_plot}
\alias{build_river_plot}
\title{Build river plots for annotations}
\usage{
build_river_plot(anno, grouping, show_labels = TRUE,
  label_pos = "center", min_link_size = 0, pad = 0.1,
  fill_group = NULL)
}
\arguments{
\item{anno}{The sample annotations to use. Must have _id, _label, and _color columns for each grouping.}

\item{show_labels}{Logical, whether or not to show labels. Default is TRUE.}

\item{label_pos}{Label position - "left", "center", or "right". Can be specified for each entry in group_by. Default = "center".}

\item{min_link_size}{Numeric, the minimum fraction of cells in either group that must be included in a link for display. Default is 0 (show all links).}

\item{pad}{The fraction of vertical space to use as padding between groups. Default = 0.1.}

\item{fill_group}{One group to use as a source for ribbon colors. Default is NULL.}

\item{group_by}{The bases to use for the river plot, from left to right.}
}
\value{
A ggplot2 plot object.
}
\description{
Build river plots for annotations
}
