% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{build_river_plot_bokeh}
\alias{build_river_plot_bokeh}
\title{River plot bokeh}
\usage{
build_river_plot_bokeh(anno, group_by, pad = 0.1, fill_group = NULL)
}
\arguments{
\item{anno}{The sample annotations to use. Must have _id, _label, and _color columns for each grouping.}

\item{group_by}{The bases to use for the river plot, from left to right.}

\item{pad}{The fraction of vertical space to use as padding between groups. Default = 0.1.}

\item{fill_group}{One group to use as a source for ribbon colors. Default is NULL.}
}
\value{
an interactive rbokeh plot object.
}
\description{
River plot bokeh
}
