% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{build_vec_pos}
\alias{build_vec_pos}
\title{Build plot positions for a character vector}
\usage{
build_vec_pos(vec, vec_name = "gene_name", sort = "none",
  axis_name = "y")
}
\arguments{
\item{vec}{a character vector (like gene names)}

\item{sort}{how to resort the vector, if necessary. Default is "none", which keeps input order. 
Options: "none","rev","random","alpha".}

\item{axis_name}{which axis to use. Default is "y"}
}
\value{
a data.frame with columns for gene_name and axis values
}
\description{
Build plot positions for a character vector
}
