% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_genes}
\alias{check_genes}
\title{Check input genes against a vector of gene names}
\usage{
check_genes(genes, gene_reference, result = "matched")
}
\arguments{
\item{genes}{An input character vector of gene symbols}

\item{gene_reference}{A reference character vector of gene symbols}

\item{result}{Whether to return a vector of matched or unmatched genes, or a list of both. Options are "matched" (default), "unmatched", or "both".}
}
\description{
Check input genes against a vector of gene names
}
