% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{data_df_to_colors}
\alias{data_df_to_colors}
\title{Convert expression data to heatmap colors for plotting}
\usage{
data_df_to_colors(df, value_cols = NULL, colorset = NULL,
  scale = "linear", per_col = FALSE, min_val = 0, max_val = NULL)
}
\arguments{
\item{df}{data.frame with sample_name and expression values}

\item{value_cols}{The value columns to convert. Default is (null), which will use all columns except sample_name}

\item{colorset}{A set of colors to use for the heatmap palette. NULL will use the default for values_to_colors()}

\item{scale}{color value scaling, passed to values_to_colors(). Default = "linear".}

\item{per_col}{Logical, whether or not to normalize the colorscale per-value, or across all values}

\item{min_val}{Minimum value for color scale. If NULL, will be automatically computed from values. Default = 0.}

\item{max_val}{Maximum value for color scale. If NULL, will be automatically computed from values. Default = 0.}
}
\description{
Convert expression data to heatmap colors for plotting
}
