% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{filter_gene_data}
\alias{filter_gene_data}
\title{Filter gene data based on annotations, genes, and groups supplied with group_order.}
\usage{
filter_gene_data(
  data,
  genes,
  anno,
  group_cols,
  group_order = NULL,
  sample_key_column = "sample_name"
)
}
\arguments{
\item{data}{a data.frame with gene expression data and a sample_name}

\item{genes}{a vector of genes to plot}

\item{anno}{a data.frame of sample annotations}

\item{group_cols}{a group_cols list generated by group_columns()}

\item{group_order}{a set of group_ids to use for the plot. Default is NULL, which ignores filtering on this parameter.}

\item{sample_key_column}{the name of the key column for sample identification. Default is "sample_name"}
}
\value{
a data.frame of data filtered based on genes, the sample_names in anno, and (if supplied) the groups in group_order.
}
\description{
Filter gene data based on annotations, genes, and groups supplied with group_order.
}
