% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{get_list_data}
\alias{get_list_data}
\title{Format data provided in list format for scrattch plots}
\usage{
get_list_data(data_list, genes, group_by, group_ids)
}
\description{
Currently only compatible with data from feather_to_list()
}
