% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{get_n_stats}
\alias{get_n_stats}
\title{Compute basic count statistics for use in generating plots}
\usage{
get_n_stats(plot_data, group_cols, genes)
}
\arguments{
\item{plot_data}{a data.frame with joined annotations and gene expression data}

\item{group_cols}{a group_cols list generated by group_columns()}

\item{genes}{a vector of genes used for the plot}
}
\value{
a list of 3 count statistics: genes, groups, and samples
}
\description{
Compute basic count statistics for use in generating plots
}
