% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{ggplot_header_labels}
\alias{ggplot_header_labels}
\title{Add header labels objects to a ggplot}
\usage{
ggplot_header_labels(p, header_labels, header_polygons = NULL, font_size)
}
\arguments{
\item{p}{The plot to add headers to}

\item{header_labels}{A data.frame of header label coordinates generated by build_header_labels()}

\item{header_polygons}{An optional data.frame of header polygons generated by build_header_polygons(). Default is NULL.}

\item{font_size}{The font size in pt to use for header label plotting.}
}
\value{
a ggplot2 object
}
\description{
Add header labels objects to a ggplot
}
