% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{ggplot_scale_bars}
\alias{ggplot_scale_bars}
\title{Add scale bar objects to a ggplot}
\usage{
ggplot_scale_bars(p, n_genes, n_samples, extent = 0.9)
}
\arguments{
\item{p}{The plot to add scale bars to}

\item{n_genes}{The number of genes in the plot}

\item{n_samples}{The number of samples in the plot}

\item{extent}{The amount of space in the plot row that each row of data uses.}
}
\value{
a ggplot2 object
}
\description{
Add scale bar objects to a ggplot
}
