% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{group_columns}
\alias{group_columns}
\title{Generate a list of anno columns based on grouping}
\usage{
group_columns(grouping)
}
\arguments{
\item{grouping}{the base of the grouping to use.}
}
\value{
a list with 3 character objects: id, label, and color
}
\description{
Generate a list of anno columns based on grouping
}
