% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{group_heatmap_plot}
\alias{group_heatmap_plot}
\title{Heatmap plots of group summary statistics}
\usage{
group_heatmap_plot(data, anno, genes, grouping, group_order = NULL,
  stat = "median", log_scale = TRUE, normalize_rows = FALSE,
  colorset = NULL, font_size = 7, label_height = 25,
  show_counts = TRUE, rotate_counts = FALSE, max_width = 10)
}
\arguments{
\item{data}{A data frame containing gene expression values. The first column should be sample_name}

\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns}

\item{genes}{A character vector containing gene symbols to be plotted.}

\item{grouping}{A character vector specifying the desc base that should be used to group cells}

\item{group_order}{Optional: Explicit specification of group order by supplying a vector of group_ids.}

\item{stat}{The statistic to apply to each group. Options are: 
\itemize{
  \item "median"
  \item "mean"
  \item "tmean" (25\% trimmed mean)
  \item "prop_gt0" (proportion of samples > 0)
  \item "prop_gt1" (proportion of samples > 1)
  \item "min"
  \item "max"
  }}

\item{log_scale}{Logical , determines if data is log scaled before plotting. Default = FALSE.}

\item{normalize_rows}{Logical, whether or not to rescale data within each row of the plot. Default = FALSE.}

\item{font_size}{numeric object, the font size (in pts) used to make the plot.}

\item{label_height}{numeric object, Percent of the plot height that should be used for the labels (0 to 100). Default is 25.}

\item{show_counts}{Logical, whether or not to display sample counts at the top of labels. Default = TRUE.}

\item{rotate_counts}{Logical, whether or not to rotate sample counts by 90 degrees. Default = FALSE.}

\item{max_width}{numeric object, percent of plot width that should be used for maximum expression values (0 to 100). Default is 10.}
}
\value{
a ggplot2 plot object
}
\description{
Heatmap plots of group summary statistics
}
