% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{group_violin_plot}
\alias{group_violin_plot}
\title{Violin plots of gene expression for grouped samples}
\usage{
group_violin_plot(
  data,
  anno,
  genes,
  grouping,
  group_order = NULL,
  log_scale = TRUE,
  font_size = 7,
  label_height = 25,
  show_counts = TRUE,
  rotate_counts = FALSE,
  max_width = 10,
  return_type = "plot"
)
}
\arguments{
\item{data}{A data frame containing gene expression values. The first column should be sample_name}

\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns}

\item{genes}{A character vector containing gene symbols to be plotted.}

\item{grouping}{A character vector specifying the desc base that should be used to group cells}

\item{group_order}{Optional: Explicit specification of group order by supplying a vector of group_ids.}

\item{log_scale}{Logical , determines if data is log scaled before plotting. Default = FALSE.}

\item{font_size}{numeric object, the font size (in pts) used to make the plot.}

\item{label_height}{numeric object, Percent of the plot height that should be used for the labels (0 to 100). Default is 25.}

\item{show_counts}{Logical, whether or not to display sample counts at the top of labels. Default = TRUE.}

\item{rotate_counts}{Logical, whether or not to rotate sample counts by 90 degrees. Default = FALSE.}

\item{max_width}{numeric object, percent of plot width that should be used for maximum expression values (0 to 100). Default is 10.}

\item{return_type}{What values to return - can be "plot", "data", or "both". Default is "plot".}
}
\value{
a ggplot2 plot object
}
\description{
Violin plots of gene expression for grouped samples
}
