% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{heatmap_legend_plot}
\alias{heatmap_legend_plot}
\title{Build a heatmap legend plot}
\usage{
heatmap_legend_plot(
  min_val = 0,
  max_val = 4,
  scale_name = "FPKM",
  colorset = c("darkblue", "dodgerblue", "gray80", "orange", "orangered")
)
}
\arguments{
\item{min_val}{numeric, the minimum value in the plot scale (default = 0)}

\item{max_val}{numeric, the maximum value in the plot scale (default = 4)}

\item{colorset}{character vector, the colors to interpolate between using colorRampPalette.}

\item{ggplot2::scale_name}{character, the name for the values displayed (default = "FPKM")}
}
\value{
a ggplot2 heatmap legend plot
}
\description{
Build a heatmap legend plot
}
