% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_group_links}
\alias{make_group_links}
\title{Make group links}
\usage{
make_group_links(anno, group_by, plot_nodes)
}
\arguments{
\item{anno}{annotation data}

\item{group_by}{grouping}

\item{plot_nodes}{output of make_plot_nodes()}
}
\description{
Make group links
}
