% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_plot_links}
\alias{make_plot_links}
\title{Generate links in plot space between nodes for a river plot}
\usage{
make_plot_links(group_links, fill = NULL)
}
\arguments{
\item{group_links}{output of make_group_links()}

\item{fill}{the group to use for fill colors}
}
\value{
a data.frame compatible with geom_ribbon() in ggplot2.
}
\description{
Generate links in plot space between nodes for a river plot
}
