% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{max_gene_vals}
\alias{max_gene_vals}
\title{Compute maximum expression values in data for a set of genes.}
\usage{
max_gene_vals(data, genes)
}
\arguments{
\item{data}{a data.frame with gene expression data}

\item{genes}{the genes to plot}
}
\value{
a numeric vector with maximum expression values named for each gene.
}
\description{
Compute maximum expression values in data for a set of genes.
}
