% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{melt_data_df}
\alias{melt_data_df}
\title{Melt a data_df to prepare for plot parameters}
\usage{
melt_data_df(data_df, grouping = "sample_name", value_cols = NULL)
}
\arguments{
\item{data_df}{a data.frame with sample_name or grouping as well as value columns}

\item{grouping}{grouping for the samples. Default = "sample_name", but if this is output from group_stats, provide the same grouping.}

\item{value_cols}{The value columns to include. If NULL (default), will automatically choose all non-grouping columns.}
}
\description{
Melt a data_df to prepare for plot parameters
}
