% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{point_size_legend_plot}
\alias{point_size_legend_plot}
\title{Build a point size legend plot}
\usage{
point_size_legend_plot(
  min_val = 0,
  max_val = 1,
  max_size = 6,
  scale_name = "Fraction of cells",
  n_sizes = 6
)
}
\arguments{
\item{min_val}{numeric, the minimum value in the plot scale (default = 0)}

\item{max_val}{numeric, the maximum value in the plot scale (default = 1)}

\item{max_size}{numeric, the max size of the point scale (default = 6, default for scale_size_area() )}

\item{scale_name}{character, the name for the values displayed (default = "Fraction of cells")}

\item{n_sizes}{numeric, the number of points in the scale from min_val to max_val, inclusive (default = 6)}
}
\value{
a ggplot2 point size legend plot
}
\description{
Build a point size legend plot
}
