% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prop_gt0}
\alias{prop_gt0}
\title{Compute the proportion/fraction of values > 0 in a numeric vector}
\usage{
prop_gt0(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Compute the proportion/fraction of values > 0 in a numeric vector
}
