% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{sample_bar_plot}
\alias{sample_bar_plot}
\title{Barplots of gene expression of individual samples}
\usage{
sample_bar_plot(data, anno, genes, grouping, group_order = NULL,
  log_scale = FALSE, font_size = 7, label_height = 25,
  label_type = "angle", max_width = 10, bg_color = "#ADCFE0")
}
\arguments{
\item{data}{A data frame containing gene expression values. The first column should be sample_name}

\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns}

\item{genes}{A character vector containing gene symbols to be plotted.}

\item{grouping}{A character vector specifying the desc base that should be used to group cells}

\item{group_order}{Optional: Explicit specification of group order by supplying a vector of group_ids.}

\item{log_scale}{Logical , determines if data is log scaled before plotting. Default = FALSE.}

\item{font_size}{numeric object, the font size (in pts) used to make the plot.}

\item{label_height}{numeric object, Percent of the plot height that should be used for the labels (0 to 100). Default is 25.}

\item{label_type}{Label shape, "angle" or "square"}

\item{max_width}{numeric object, percent of plot width that should be used for maximum expression values (0 to 100). Default is 10.}

\item{bg_color}{plot background color. Default is a light blue ("#ADCFE0")}
}
\value{
a ggplot2 plot object
}
\description{
This function will generate plots similar to those in Figure 3a-c of Tasic, et al. (2016).
}
