% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{sample_fire_plot}
\alias{sample_fire_plot}
\title{Fire Heatmaps of gene expression for individual samples}
\usage{
sample_fire_plot(data, anno, genes, grouping, group_order = NULL,
  log_scale = TRUE, normalize_rows = FALSE, colorset = NULL,
  top_values = "lowest", font_size = 7, label_height = 25,
  max_width = 10)
}
\arguments{
\item{data}{A data frame containing gene expression values. The first column should be sample_name}

\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns}

\item{genes}{A character vector containing gene symbols to be plotted.}

\item{grouping}{A character vector specifying the desc base that should be used to group cells}

\item{group_order}{Optional: Explicit specification of group order by supplying a vector of group_ids.}

\item{log_scale}{Logical , determines if data is log scaled before plotting. Default = FALSE.}

\item{normalize_rows}{Logical, determines if heatmaps will be normalized for each gene. Default = FALSE.}

\item{font_size}{numeric object, the font size (in pts) used to make the plot.}

\item{label_height}{numeric object, Percent of the plot height that should be used for the labels (0 to 100). Default is 25.}

\item{max_width}{numeric object, percent of plot width that should be used for maximum expression values (0 to 100). Default is 10.}
}
\value{
a ggplot2 plot object
}
\description{
Fire Heatmaps of gene expression for individual samples
}
