% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{scale_gene_data}
\alias{scale_gene_data}
\title{Scale gene data columns}
\usage{
scale_gene_data(data, genes, scale_type = "log10")
}
\arguments{
\item{data}{a data.frame with gene expression data}

\item{genes}{the genes to plot}

\item{scale_type}{the type of scaling to supply. Currently only supports "log10" (default), and "linear" (no rescaling).}
}
\value{
a data.frame with columns matching genes rescaled based on scale_type.
}
\description{
Scale gene data columns
}
