\name{scrattch.vis-package}
\alias{scrattch.vis-package}
\alias{scrattch.vis}
\docType{package}
\title{
\packageTitle{scrattch.vis}
}
\description{
\packageDescription{scrattch.vis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scrattch.vis}
\packageIndices{scrattch.vis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{scrattch.vis}

Maintainer: \packageMaintainer{scrattch.vis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
