% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sigline}
\alias{sigline}
\title{Build a sigmoidal line using sigfun}
\usage{
sigline(x = 0, xend = 1, y = 0, yend = 1, steps = 50,
  sigfun = "erf", sigx = 1.5)
}
\arguments{
\item{x}{starting x position. Default = 0}

\item{xend}{ending x position. Default = 1}

\item{y}{starting y position. Default = 0}

\item{yend}{ending y position. Default = 1}

\item{steps}{Number of points in the line. Default = 50}

\item{sigfun}{The sigmoidal function to use. Currently only supports "erf"}

\item{sigx}{Will generate values for -sigx to +sigx using sigfun, then rescale to x and y params.}
}
\value{
a data.frame with x and y coordinates for the sigmoidal line
}
\description{
Build a sigmoidal line using sigfun
}
