% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sigribbon}
\alias{sigribbon}
\title{Expand a sigmoidal line to a ribbon by adding ymin}
\usage{
sigribbon(sigline, height, from = "top")
}
\arguments{
\item{height}{The desired height of the ribbon}

\item{from}{Whether the line should be treated as the "top", "bot"(tom), or "mid"(dle) of the ribbon}

\item{signline}{a data.frame with x and y coordinates for a sigmoidal line}
}
\value{
a data.frame with coordinates x, y, and ymin for the ribbon.
}
\description{
Expand a sigmoidal line to a ribbon by adding ymin
}
