% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tmean}
\alias{tmean}
\title{Compute Trimmed Mean of a numeric vector}
\usage{
tmean(x, trim = 0.25)
}
\arguments{
\item{x}{A numeric vector}

\item{trim}{the amount to trim}
}
\description{
This is a simple wrapper around mean(x, trim)
}
