% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{build_river_plot_bokeh}
\alias{build_river_plot_bokeh}
\title{River plot bokeh}
\usage{
build_river_plot_bokeh(anno, group_by, pad = 0.1, fill_group = NULL)
}
\arguments{
\item{anno}{}

\item{group_by}{}

\item{pad}{}

\item{fill_group}{}
}
\description{
River plot bokeh
}
