% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dend_anno.R
\name{dend_anno_barplot}
\alias{dend_anno_barplot}
\title{dend_anno_barplot}
\usage{
dend_anno_barplot(
  anno,
  dend,
  section_wedges = NULL,
  bar_variables = NULL,
  nr_hist = TRUE,
  panel_width = 0.2
)
}
\arguments{
\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns. Requires cluster_id which needs to be sequential in order of the dendrogram.}

\item{dend}{Dendrogram object.}

\item{section_wedges}{Default is NULL. Use annotation to separate  Can be used to generate lines between sections to divide leaves of dendrogram, e.g. separating subclass, class.}

\item{bar_variables}{base name of variables to be represented as bargraphs below dendrogram. Annotation variables need to be represented as \_id, \_label, \_color in anno.}

\item{nr_hist}{plotting of cluster size below dend. Default is TRUE.}

\item{return_type}{What values to return - can be "plot", "data", or "both". Default is "plot".}
}
\description{
dend_anno_barplot
}
\examples{
# First, load the example data
load("data/dend_anno_example/dend.rda")
load("data/dend_anno_example/anno.df.rda")

# Then, run the function with the loaded data
dend.plot <- dend_anno_barplot(
  anno = anno.df,
  dend = dend,
  section_wedges = "class",
  bar_variables = c("sex", "batch"),
  nr_hist = TRUE
)
}
