% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dend_anno.R
\name{dend_anno_dotplot}
\alias{dend_anno_dotplot}
\title{dend_anno_dotplot}
\usage{
dend_dotplot <- dend_anno_dotplot(plot_anno=anno.df, dend=dend, y_group="genotype", c_group="genotype", denom=NULL)
}
\arguments{
\item{dend}{Dendrogram object. Needs to be labeled with cluster_label}

\item{y_group}{base name of variable to be represented on y-axis of dotplot below dendrogram. Annotation variable needs to be represented as \_id, \_label, \_color in anno.}

\item{c_group}{base name of variables to color the dotplot by. Annotation variables need to be represented as \_id, \_label, \_color in anno.}

\item{prefix}{prefix that groups cluster_id/_label/_color. Used in case of mapping to multiple taxonomies. Default is NULL.}

\item{denom}{When calculating fraction of either x- or y-axis variables. Input can be "cluster" or y_group. Default is NULL.}

\item{plot_anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns. Requires cluster_id which needs to be sequential in order of the dendrogram.}
}
\description{
dend_anno_dotplot
}
