% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_group_nodes}
\alias{make_group_nodes}
\title{Make group nodes for categorical data}
\usage{
make_group_nodes(anno, group_by, value_col, xpos = NULL)
}
\arguments{
\item{anno}{sample annotations}

\item{group_by}{which columns of anno to use for sample grouping (should be at least 2). Group must have "_id", "_label", "_color" columns.}
}
\description{
Make group nodes for categorical data
}
