% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{mat_to_data_df}
\alias{mat_to_data_df}
\title{Convert a matrix to a data.frame for plotting}
\usage{
mat_to_data_df(mat, cols_are = "gene_names")
}
\arguments{
\item{mat}{a matrix or sparse matrix dgCMatrix object from the Matrix package.}

\item{cols_are}{whether columns are "gene_names" or "sample_names"}
}
\value{
dataframe
}
\description{
Convert a matrix to a data.frame for plotting
}
