% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{sample_bar_plot}
\alias{sample_bar_plot}
\title{Barplots of gene expression of individual cells}
\usage{
sample_bar_plot(data, anno, genes, group_by, sort = F, fontsize = 7,
  labelheight = 25, labeltype = "angle", max_width = 10,
  bgcolor = "#ADCFE0")
}
\arguments{
\item{data}{A data frame containing gene expression values. The first column should be sample_name}

\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns}

\item{genes}{A character vector containing gene symbols to be plotted.}

\item{group_by}{A character vector specifying the desc base that should be used to group cells}

\item{sort}{Logical object, determines if cells will be sorted within their clusters (this means that each column will no longer represent a single cell)}

\item{fontsize}{numeric object, the font size (in pts) used to make the plot.}

\item{labelheight}{numeric object, Percent of the plot height that should be used for the labels (0 to 100).}

\item{logscale}{Logical object, determines if data is log scaled before plotting.}
}
\value{
a ggplot2 plot object
}
\description{
This function will generate plots similar to those in Figure 3a-c of Tasic, et al. (2015).
Warning: this is currently only able to work with internally-supplied datasets (v1_data and v1_anno).
Extension to user-supplied datasets will come soon.
See Examples for sensible PDF output options.
}
\examples{

my_genes <- c("Ercc6","Ercc8","Trp53","Pgbd5")
my_clusters <- c(1,5,9,10,24,37)
my_barcell_plot <- barcell_plot(my_genes,my_clusters,sort=T)

ggsave("plot_output.pdf",my_barcell_plot,height=0.2*length(my_genes)+2,width=4)
}
